Deployment instructions and operations
This initial commit contains all the necessary bits to make a
kubernetes cluster and install codex instances.

This was copied from a previously private repo where I developed the
first iteration of this without concern for security. Thus I committed
some secrets in that git repository. I've made changes since then to
produce the secrets and other configuration on the commandline rather
than in static yaml files.

This repo is now secret safe. All sensative information is either not
committed or pushed into the respective cluster on creation.